modification: exception handling
creating a exception handler decorator to print an exception message
to cli.
